Change time variables in class 'VHSTape' from 'double' to 'int' to represent lenght of the movie in minutes

+++ VHSTape.cs
@@ -5,18 +5,18 @@ namespace Activity4._4._3
     class VHSTape
     {
         private readonly string _name;
-        private double _played;
-        private readonly double _length;
+        private int _played;
+        private readonly int _length;
         public bool Rented {set; get;}
 
-        public VHSTape(string name, double length)
+        public VHSTape(string name, int length)
         {
             this._name = name;
             this._length = length;
             this.Rented = false;
-            this._played = 0.0;
+            this._played = 0;
         }
-        public void Play(double time)
+        public void Play(int time)
         {
             this._played += time;
             if(this._played > this._length) 
@@ -24,12 +24,12 @@ namespace Activity4._4._3
                 this._played = this._length;
             }
         }
-        public void Rewind(double time)
+        public void Rewind(int time)
         {
             this._played -= time;
-            if(this._played < 0.0)
+            if(this._played < 0)
             {
-                this._played = 0.0;
+                this._played = 0;
             }
         }
         public string GetName()
